feat: KontextInterval documentation and merge with Kontext
